resolve_bet and refund_bet functions done

+++ programs/dice-game/src/state/mod.rs
@@ -3,7 +3,7 @@ use anchor_lang::prelude::*;
 #[account]
 pub struct Bet {
     pub player: Pubkey,
-    pub seed: u64,
+    pub seed: u128,
     pub amount: u64,
     pub roll: u8,
     pub slot: u64,
